refactor(valibot): extract number helper functions to separate file
